# openapi.yaml
openapi: 3.0.3
info:
  title: My Business
  description: |-
    An application that helps small business companies with their daily transactions.
  version: 0.0.1
...
servers:
  - url: https://localhost:8080/api
...
paths:
...
  /users:
    post:
      description: Creates a new user.
      summary: Creates a new user.
      operationId: createUser
      tags:
      - users
      requestBody:
        description: New user details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserModel'
        required: true
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdModel'
        400:
          description: Not valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
...
components:
  schemas:
    IdModel:
      type: object
      properties:
        id: 
          type: integer
          format: int64
          example: 22
...
    UserModel:
      type: object
      properties:
        username:
          type: string
          example: username
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string  
          example: john@email.com
        phone:
          type: string
          example: '12345'
    NewUserModel:
      allOf:
        - $ref: "#/components/schemas/UserModel"
        - type: object
          properties:
            password:
              type: string
              example: '12345'
    ErrorModel:
      type: object
      properties:
        message: 
          type: string
          example: Not valid request
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT